pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "fastapi>=0.115.12",
    "llama-parse>=0.6.12",
    "mistralai>=1.7.0",
    "openai>=1.75.0",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.1",
]
